+++ middlewares/add-filter.mjs
@@ -3,6 +3,7 @@ export default (id, mongoPath) => async (ctx, next) => {
   const filters = ctx.filters || {};
   if (queryFilters[id]) {
     filters[mongoPath] = queryFilters[id];
+    delete ctx.query.filters[id];
   }
   ctx.filters = filters;
   await next();