+++ tests/routes/jwt.test.mjs
@@ -13,7 +13,7 @@ describe('Test the JWT route', () => {
         .post('/jwt/generate')
         .send({
           email: 'foo@bar.com',
-          password: 'foobar',
+          password: 'foobarbaz',
         });
       expect(statusCode).to.equal(403);
     });
@@ -23,13 +23,13 @@ describe('Test the JWT route', () => {
         .post('/jwt/generate')
         .send({
           email: 'foo@bar.com',
-          password: 'foobar',
+          password: 'foobarbaz',
         });
       expect(statusCode).to.equal(403);
     });
 
     it('It should return a valid token on generate', async () => {
-      const PASSWORD = 'foobar';
+      const PASSWORD = 'foobarbaz';
       const user = await createDummyUser({ password: PASSWORD });
       try {
         const { body: { token }, statusCode } = await request()
@@ -45,7 +45,7 @@ describe('Test the JWT route', () => {
         )
           .to.deep.include({ email: user.email, id: user.id });
       } finally {
-        await user.remove();
+        await user.deleteOne();
       }
     });
   });
@@ -76,12 +76,12 @@ describe('Test the JWT route', () => {
           .set('Authorization', `Bearer ${token}`);
         expect(statusCode).to.equal(404);
       } finally {
-        await user.remove();
+        await user.deleteOne();
       }
     });
 
     it('It should return a valid new token on renew', async () => {
-      const PASSWORD = 'foobar';
+      const PASSWORD = 'foobarbaz';
       const user = await createDummyUser({ password: PASSWORD });
       const authToken = user.emitJWT();
 
@@ -96,14 +96,14 @@ describe('Test the JWT route', () => {
         )
           .to.deep.include({ email: user.email, id: user.id });
       } finally {
-        await user.remove();
+        await user.deleteOne();
       }
     });
   });
 
   describe('Test the GET user route', () => {
     it('It should return a 404 error on get user when user not exists', async () => {
-      const PASSWORD = 'foobar';
+      const PASSWORD = 'foobarbaz';
       const user = new User(generateDummyUser({ password: PASSWORD }));
       const token = user.emitJWT();
 
@@ -113,12 +113,12 @@ describe('Test the JWT route', () => {
           .set('Authorization', `Bearer ${token}`);
         expect(statusCode).to.equal(404);
       } finally {
-        await user.remove();
+        await user.deleteOne();
       }
     });
 
     it('It should return the user', async () => {
-      const PASSWORD = 'foobar';
+      const PASSWORD = 'foobarbaz';
       const u = await createDummyUser({ password: PASSWORD });
       const authToken = u.emitJWT();
 
@@ -131,7 +131,7 @@ describe('Test the JWT route', () => {
         expect(user)
           .to.deep.equal({ email: u.email, id: u.id });
       } finally {
-        await u.remove();
+        await u.deleteOne();
       }
     });
   });